requirements.txt
langchain
langchain_community
langchain-openai
streamlit
openai
python-dotenv
faiss-cpu
langchainhub
sentence_transformers